Remove field_override, move functionality elsewhere
